Offload the responsibilities to the coordinator
